+++ backend/src/schema/budget.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 #[serde(rename_all = "camelCase")]
 pub struct CreateBudgetSchema {
     pub name: String,
-    pub amount: f32,
+    pub amount: i64,
     pub start_date: chrono::NaiveDate,
     pub end_date: chrono::NaiveDate,
 }
