+++ backend/src/schema/expense.rs
@@ -4,8 +4,8 @@ use serde::Deserialize;
 #[serde(rename_all = "camelCase")]
 pub struct CreateExpenseSchema {
     pub name: String,
-    pub amount: f32,
+    pub amount: i64,
     pub date: chrono::NaiveDate,
     pub description: Option<String>,
-    pub category_id: Option<u64>,
+    pub category_id: Option<i32>,
 }
